refactor(reduce): merge max and min into one instruction
The two files were 285 lines each and differed by one doc word once the
direction words are normalized. They are now one Extremum carrying the
ValueOrder it ranks by, which also dissolves PackedExtremum: that type
existed only to pair a Packing with methods the instruction can own, and
the order it packed in was the order the instruction already knew.

ReduceOperation loses a variant with them, so ten pairs of delegating match
arms become ten arms. ReduceOperationConfig is untouched, since that is what
callers outside the crate name.

The instruction's config has to carry the order, because from_config cannot
otherwise tell the extremes apart, which makes ValueOrder public.

Emitted code is unchanged: 643 kernels untouched by the rename are
byte-identical, and the 50 max and 50 min kernels match the descending and
ascending ones they became.
