+++ crates/cubek-reduce/src/components/instructions/mixed.rs
@@ -1,10 +1,10 @@
 use super::{
-    All, Any, Max, MaxAbs, Mean, Min, Prod, ReduceFamily, ReduceInstruction, ReduceRequirements,
+    All, Any, Extremum, MaxAbs, Mean, Prod, ReduceFamily, ReduceInstruction, ReduceRequirements,
     SharedAccumulator, Sum,
 };
 use crate::components::instructions::{
     Accumulator, AccumulatorExpand, AccumulatorFormat, Item, Packed, ReduceOutputMode,
-    SharedAccumulatorKind, SlotCount, TopK,
+    SharedAccumulatorKind, SlotCount, TopK, ValueOrder,
 };
 use crate::{
     ReduceDtypes,
@@ -25,8 +25,7 @@ pub enum ReduceOperation {
     Prod(Prod),
     Mean(Mean),
     MaxAbs(MaxAbs),
-    Max(Max),
-    Min(Min),
+    Extremum(Extremum),
     TopK(TopK),
     Any(Any),
     All(All),
@@ -290,8 +289,9 @@ impl<P: ReducePrecision> ReduceInstruction<P> for ReduceOperation {
                 <MaxAbs as ReduceInstruction<P>>::requirements(max_abs)
             }
             ReduceOperation::TopK(topk) => <TopK as ReduceInstruction<P>>::requirements(topk),
-            ReduceOperation::Max(max) => <Max as ReduceInstruction<P>>::requirements(max),
-            ReduceOperation::Min(min) => <Min as ReduceInstruction<P>>::requirements(min),
+            ReduceOperation::Extremum(extremum) => {
+                <Extremum as ReduceInstruction<P>>::requirements(extremum)
+            }
             ReduceOperation::Any(any) => <Any as ReduceInstruction<P>>::requirements(any),
             ReduceOperation::All(all) => <All as ReduceInstruction<P>>::requirements(all),
         }
@@ -305,8 +305,9 @@ impl<P: ReducePrecision> ReduceInstruction<P> for ReduceOperation {
             ReduceOperation::MaxAbs(maxabs) => {
                 <MaxAbs as ReduceInstruction<P>>::accumulator_format(maxabs)
             }
-            ReduceOperation::Max(max) => <Max as ReduceInstruction<P>>::accumulator_format(max),
-            ReduceOperation::Min(min) => <Min as ReduceInstruction<P>>::accumulator_format(min),
+            ReduceOperation::Extremum(extremum) => {
+                <Extremum as ReduceInstruction<P>>::accumulator_format(extremum)
+            }
             ReduceOperation::TopK(topk) => <TopK as ReduceInstruction<P>>::accumulator_format(topk),
             ReduceOperation::Any(any) => <Any as ReduceInstruction<P>>::accumulator_format(any),
             ReduceOperation::All(all) => <All as ReduceInstruction<P>>::accumulator_format(all),
@@ -319,20 +320,24 @@ impl<P: ReducePrecision> ReduceInstruction<P> for ReduceOperation {
             ReduceOperationConfig::Prod => ReduceOperation::new_Prod(Prod {}),
             ReduceOperationConfig::Mean => ReduceOperation::new_Mean(Mean { sum: Sum {} }),
             ReduceOperationConfig::MaxAbs => ReduceOperation::new_MaxAbs(MaxAbs {}),
-            ReduceOperationConfig::ArgMax => ReduceOperation::new_Max(Max {
+            ReduceOperationConfig::ArgMax => ReduceOperation::new_Extremum(Extremum {
+                order: ValueOrder::Descending,
                 output: ReduceOutputMode::Indices,
             }),
-            ReduceOperationConfig::ArgMin => ReduceOperation::new_Min(Min {
+            ReduceOperationConfig::ArgMin => ReduceOperation::new_Extremum(Extremum {
+                order: ValueOrder::Ascending,
                 output: ReduceOutputMode::Indices,
             }),
             ReduceOperationConfig::ArgTopK(k) => ReduceOperation::new_TopK(TopK {
                 k,
                 output: ReduceOutputMode::Indices,
             }),
-            ReduceOperationConfig::Max => ReduceOperation::new_Max(Max {
+            ReduceOperationConfig::Max => ReduceOperation::new_Extremum(Extremum {
+                order: ValueOrder::Descending,
                 output: ReduceOutputMode::Values,
             }),
-            ReduceOperationConfig::Min => ReduceOperation::new_Min(Min {
+            ReduceOperationConfig::Min => ReduceOperation::new_Extremum(Extremum {
+                order: ValueOrder::Ascending,
                 output: ReduceOutputMode::Values,
             }),
             ReduceOperationConfig::TopK(k) => ReduceOperation::new_TopK(TopK {
@@ -350,8 +355,9 @@ impl<P: ReducePrecision> ReduceInstruction<P> for ReduceOperation {
             ReduceOperation::Prod(prod) => <Prod as ReduceInstruction<P>>::null_input(prod),
             ReduceOperation::Mean(mean) => <Mean as ReduceInstruction<P>>::null_input(mean),
             ReduceOperation::MaxAbs(maxabs) => <MaxAbs as ReduceInstruction<P>>::null_input(maxabs),
-            ReduceOperation::Max(max) => <Max as ReduceInstruction<P>>::null_input(max),
-            ReduceOperation::Min(min) => <Min as ReduceInstruction<P>>::null_input(min),
+            ReduceOperation::Extremum(extremum) => {
+                <Extremum as ReduceInstruction<P>>::null_input(extremum)
+            }
             ReduceOperation::TopK(topk) => <TopK as ReduceInstruction<P>>::null_input(topk),
             ReduceOperation::Any(any) => <Any as ReduceInstruction<P>>::null_input(any),
             ReduceOperation::All(all) => <All as ReduceInstruction<P>>::null_input(all),
@@ -366,8 +372,9 @@ impl<P: ReducePrecision> ReduceInstruction<P> for ReduceOperation {
             ReduceOperation::MaxAbs(maxabs) => {
                 <MaxAbs as ReduceInstruction<P>>::null_accumulator(maxabs)
             }
-            ReduceOperation::Max(max) => <Max as ReduceInstruction<P>>::null_accumulator(max),
-            ReduceOperation::Min(min) => <Min as ReduceInstruction<P>>::null_accumulator(min),
+            ReduceOperation::Extremum(extremum) => {
+                <Extremum as ReduceInstruction<P>>::null_accumulator(extremum)
+            }
             ReduceOperation::TopK(topk) => <TopK as ReduceInstruction<P>>::null_accumulator(topk),
             ReduceOperation::Any(any) => <Any as ReduceInstruction<P>>::null_accumulator(any),
             ReduceOperation::All(all) => <All as ReduceInstruction<P>>::null_accumulator(all),
@@ -393,11 +400,8 @@ impl<P: ReducePrecision> ReduceInstruction<P> for ReduceOperation {
             ReduceOperation::MaxAbs(maxabs) => {
                 <MaxAbs as ReduceInstruction<P>>::reduce(maxabs, accumulator, item, reduce_step)
             }
-            ReduceOperation::Max(max) => {
-                <Max as ReduceInstruction<P>>::reduce(max, accumulator, item, reduce_step)
-            }
-            ReduceOperation::Min(min) => {
-                <Min as ReduceInstruction<P>>::reduce(min, accumulator, item, reduce_step)
+            ReduceOperation::Extremum(extremum) => {
+                <Extremum as ReduceInstruction<P>>::reduce(extremum, accumulator, item, reduce_step)
             }
             ReduceOperation::TopK(topk) => {
                 <TopK as ReduceInstruction<P>>::reduce(topk, accumulator, item, reduce_step)
@@ -425,11 +429,8 @@ impl<P: ReducePrecision> ReduceInstruction<P> for ReduceOperation {
             ReduceOperation::MaxAbs(max_abs) => {
                 <MaxAbs as ReduceInstruction<P>>::plane_reduce_inplace(max_abs, accumulator)
             }
-            ReduceOperation::Max(max) => {
-                <Max as ReduceInstruction<P>>::plane_reduce_inplace(max, accumulator)
-            }
-            ReduceOperation::Min(min) => {
-                <Min as ReduceInstruction<P>>::plane_reduce_inplace(min, accumulator)
+            ReduceOperation::Extremum(extremum) => {
+                <Extremum as ReduceInstruction<P>>::plane_reduce_inplace(extremum, accumulator)
             }
             ReduceOperation::TopK(topk) => {
                 <TopK as ReduceInstruction<P>>::plane_reduce_inplace(topk, accumulator)
@@ -457,11 +458,8 @@ impl<P: ReducePrecision> ReduceInstruction<P> for ReduceOperation {
             ReduceOperation::MaxAbs(maxabs) => {
                 <MaxAbs as ReduceInstruction<P>>::fuse_accumulators(maxabs, accumulator, other)
             }
-            ReduceOperation::Max(max) => {
-                <Max as ReduceInstruction<P>>::fuse_accumulators(max, accumulator, other)
-            }
-            ReduceOperation::Min(min) => {
-                <Min as ReduceInstruction<P>>::fuse_accumulators(min, accumulator, other)
+            ReduceOperation::Extremum(extremum) => {
+                <Extremum as ReduceInstruction<P>>::fuse_accumulators(extremum, accumulator, other)
             }
             ReduceOperation::TopK(topk) => {
                 <TopK as ReduceInstruction<P>>::fuse_accumulators(topk, accumulator, other)
@@ -483,8 +481,9 @@ impl<P: ReducePrecision> ReduceInstruction<P> for ReduceOperation {
             ReduceOperation::MaxAbs(maxabs) => {
                 <MaxAbs as ReduceInstruction<P>>::output_mode(maxabs)
             }
-            ReduceOperation::Max(max) => <Max as ReduceInstruction<P>>::output_mode(max),
-            ReduceOperation::Min(min) => <Min as ReduceInstruction<P>>::output_mode(min),
+            ReduceOperation::Extremum(extremum) => {
+                <Extremum as ReduceInstruction<P>>::output_mode(extremum)
+            }
             ReduceOperation::TopK(topk) => <TopK as ReduceInstruction<P>>::output_mode(topk),
             ReduceOperation::Any(any) => <Any as ReduceInstruction<P>>::output_mode(any),
             ReduceOperation::All(all) => <All as ReduceInstruction<P>>::output_mode(all),
@@ -516,14 +515,13 @@ impl<P: ReducePrecision> ReduceInstruction<P> for ReduceOperation {
                     shape_axis_reduce,
                 )
             }
-            ReduceOperation::Max(max) => <Max as ReduceInstruction<P>>::to_output_parallel::<
-                Out,
-                Idx,
-            >(max, accumulator, shape_axis_reduce),
-            ReduceOperation::Min(min) => <Min as ReduceInstruction<P>>::to_output_parallel::<
-                Out,
-                Idx,
-            >(min, accumulator, shape_axis_reduce),
+            ReduceOperation::Extremum(extremum) => {
+                <Extremum as ReduceInstruction<P>>::to_output_parallel::<Out, Idx>(
+                    extremum,
+                    accumulator,
+                    shape_axis_reduce,
+                )
+            }
             ReduceOperation::TopK(topk) => <TopK as ReduceInstruction<P>>::to_output_parallel::<
                 Out,
                 Idx,
@@ -570,14 +568,13 @@ impl<P: ReducePrecision> ReduceInstruction<P> for ReduceOperation {
                     shape_axis_reduce,
                 )
             }
-            ReduceOperation::Max(max) => <Max as ReduceInstruction<P>>::to_output_perpendicular::<
-                Out,
-                Idx,
-            >(max, accumulator, shape_axis_reduce),
-            ReduceOperation::Min(min) => <Min as ReduceInstruction<P>>::to_output_perpendicular::<
-                Out,
-                Idx,
-            >(min, accumulator, shape_axis_reduce),
+            ReduceOperation::Extremum(extremum) => {
+                <Extremum as ReduceInstruction<P>>::to_output_perpendicular::<Out, Idx>(
+                    extremum,
+                    accumulator,
+                    shape_axis_reduce,
+                )
+            }
             ReduceOperation::TopK(topk) => {
                 <TopK as ReduceInstruction<P>>::to_output_perpendicular::<Out, Idx>(
                     topk,
