+++ crates/cubek-reduce/src/components/instructions/mod.rs
@@ -1,13 +1,10 @@
 mod all;
 mod any;
 mod base;
-mod extrema;
-mod max;
+mod extremum;
 mod maxabs;
 mod mean;
-mod min;
 mod mixed;
-mod packed_extremum;
 mod packing;
 mod prod;
 mod sum;
@@ -17,13 +14,11 @@ mod utils;
 pub use all::*;
 pub use any::*;
 pub use base::*;
-pub(crate) use extrema::*;
-pub use max::*;
+pub use extremum::*;
 pub use maxabs::*;
 pub use mean::*;
-pub use min::*;
 pub use mixed::*;
-pub(crate) use packed_extremum::*;
+pub use packing::ValueOrder;
 pub(crate) use packing::*;
 pub use prod::*;
 pub use sum::*;
